+++ ZakYip.Singulation.Drivers/Abstractions/Events/AxisSpeedFeedbackEventArgs.cs
@@ -19,25 +19,25 @@ namespace ZakYip.Singulation.Drivers.Abstractions.Events {
         /// 实际转速（rpm）。
         /// <para>数学：rpm = rev/s × 60。</para>
         /// </summary>
-        public double Rpm { get; }
+        public decimal Rpm { get; }
 
         /// <summary>
         /// 实际线速度（m/s）。
         /// <para>数学：v = (rpm / 60) × (π·D)。</para>
         /// </summary>
-        public double SpeedMps { get; }
+        public decimal SpeedMps { get; }
 
         /// <summary>
         /// 实际脉冲频率（pps, pulses per second）。
         /// <para>数学：pps = (rpm / 60) × PPR。</para>
         /// </summary>
-        public double PulsesPerSec { get; }
+        public decimal PulsesPerSec { get; }
 
         /// <summary>时间戳（UTC）。</summary>
         public DateTime TimestampUtc { get; }
 
         public AxisSpeedFeedbackEventArgs(
-            AxisId axis, double rpm, double speedMps, double pulsesPerSec, DateTime timestampUtc) {
+            AxisId axis, decimal rpm, decimal speedMps, decimal pulsesPerSec, DateTime timestampUtc) {
             Axis = axis;
             Rpm = rpm;
             SpeedMps = speedMps;
